+++ template/src/hb_helpers.js
@@ -16,7 +16,7 @@ import Handlebars from 'handlebars';
 import { __ } from './locales/locale';
 import $ from 'jquery';
 import { body2json } from './jsonifier';
-import DiffMatchPatch from 'diff-match-patch';
+import DiffMatchPatch from './diff_match_patch';
 
 // this will register all helpers
 export function register () {
@@ -274,9 +274,10 @@ export function register () {
 
       if (!compare) { return source; }
 
-      const d = diffMatchPatch.diff_main(stripHtml(compare), stripHtml(source));
-      diffMatchPatch.diff_cleanupSemantic(d);
-      ds = diffMatchPatch.diff_prettyHtml(d);
+      const diffMatchPatch = new DiffMatchPatch();
+      const d = diffMatchPatch.diffMain(compare, source);
+      diffMatchPatch.diffCleanupSemantic(d);
+      ds = diffMatchPatch.diffPrettyHtml(d);
       ds = ds.replace(/&para;/gm, '');
     }
     if (options === 'nl2br') { ds = _handlebarsNewlineToBreak(ds); }
@@ -344,47 +345,5 @@ export function register () {
       ret = ret + options.fn(dataList[index]);
     }
     return ret;
-  }
-
-  // FIXME this one is needed to pass elint check
-  const DiffMatchPatchWrapper = DiffMatchPatch.diff_match_patch;
-  const diffMatchPatch = new DiffMatchPatchWrapper();
-
-  /**
-   * Overwrite Colors
-   */
-  diffMatchPatch.diff_prettyHtml = function (diffs) {
-    const html = [];
-    const patternAmp = /&/g;
-    const patternLt = /</g;
-    const patternGt = />/g;
-    const patternPara = /\n/g;
-    for (let x = 0; x < diffs.length; x++) {
-      const op = diffs[x][0]; // Operation (insert, delete, equal)
-      const data = diffs[x][1]; // Text of change.
-      const text = data.replace(patternAmp, '&amp;').replace(patternLt, '&lt;')
-        .replace(patternGt, '&gt;').replace(patternPara, '&para;<br>');
-      switch (op) {
-        case DiffMatchPatch.DIFF_INSERT:
-          html[x] = '<ins>' + text + '</ins>';
-          break;
-        case DiffMatchPatch.DIFF_DELETE:
-          html[x] = '<del>' + text + '</del>';
-          break;
-        case DiffMatchPatch.DIFF_EQUAL:
-          html[x] = '<span>' + text + '</span>';
-          break;
-      }
-    }
-    return html.join('');
-  };
-
-  /**
-     * Fixes html after comparison (#506, #538, #616, #825)
-  */
-  function stripHtml (html) {
-    const div = document.createElement('div');
-    div.innerHTML = html;
-    return div.textContent || div.innerText || '';
   }
 }
